// apps/mcom_mallAPI/src/resources/offer/offer.enum.ts
export enum RewardCouponType {
  FIXED_CART_DISCOUNT = 'FIXED_CART_DISCOUNT',
  PERCENTAGE_DISCOUNT = 'PERCENTAGE_DISCOUNT',
  FREE_PRODUCTS = 'FREE_PRODUCTS',
  BONUS_POINTS = 'BONUS_POINTS',
}

export enum OfferScope {
  ALL_LISTINGS = 'ALL_LISTINGS',
  SPECIFIC_LISTINGS = 'SPECIFIC_LISTINGS',
  SPECIFIC_PRODUCTS = 'SPECIFIC_PRODUCTS',
}